+++ packages/i18n/src/t/translation-renderer.ts
@@ -8,12 +8,12 @@ import {
   renderer,
   IObserverLocator,
   IsBindingBehavior,
-  LifecycleFlags,
   IHydratableController,
   AttrSyntax,
   IPlatform,
   IAttrMapper,
   ICommandBuildInfo,
+  CustomExpression,
 } from '@aurelia/runtime-html';
 
 import type {
@@ -46,7 +46,7 @@ export class TranslationBindingInstruction {
 export class TranslationBindingCommand implements BindingCommandInstance {
   public readonly bindingType: BindingType.CustomCommand = BindingType.CustomCommand;
 
-  public static get inject() { return [IAttrMapper]; }
+  public static inject = [IAttrMapper];
   public constructor(private readonly m: IAttrMapper) {}
 
   public build(info: ICommandBuildInfo): TranslationBindingInstruction {
@@ -59,7 +59,7 @@ export class TranslationBindingCommand implements BindingCommandInstance {
     } else {
       target = info.bindable.property;
     }
-    return new TranslationBindingInstruction(info.expr as IsBindingBehavior, target);
+    return new TranslationBindingInstruction(new CustomExpression(info.attr.rawValue) as IsBindingBehavior, target);
   }
 }
 
@@ -72,7 +72,6 @@ export class TranslationBindingRenderer implements IRenderer {
   ) { }
 
   public render(
-    f: LifecycleFlags,
     renderingCtrl: IHydratableController,
     target: HTMLElement,
     instruction: CallBindingInstruction,
@@ -115,8 +114,11 @@ export class TranslationBindBindingInstruction {
 export class TranslationBindBindingCommand implements BindingCommandInstance {
   public readonly bindingType: BindingType.BindCommand = BindingType.BindCommand;
 
-  public static get inject() { return [IAttrMapper]; }
-  public constructor(private readonly m: IAttrMapper) {}
+  public static inject = [IAttrMapper, IExpressionParser];
+  public constructor(
+    private readonly m: IAttrMapper,
+    private readonly xp: IExpressionParser,
+  ) {}
 
   public build(info: ICommandBuildInfo): TranslationBindingInstruction {
     let target: string;
@@ -128,7 +130,7 @@ export class TranslationBindBindingCommand implements BindingCommandInstance {
     } else {
       target = info.bindable.property;
     }
-    return new TranslationBindBindingInstruction(info.expr as IsBindingBehavior, target);
+    return new TranslationBindBindingInstruction(this.xp.parse(info.attr.rawValue, BindingType.BindCommand), target);
   }
 }
 
@@ -141,7 +143,6 @@ export class TranslationBindBindingRenderer implements IRenderer {
   ) { }
 
   public render(
-    f: LifecycleFlags,
     renderingCtrl: IHydratableController,
     target: HTMLElement,
     instruction: CallBindingInstruction,