+++ packages/runtime-html/src/configuration.ts
@@ -68,9 +68,9 @@ import {
   FulfilledTemplateController,
   RejectedTemplateController,
   // TODO: activate after the attribute parser and/or interpreter such that for `t`, `then` is not picked up.
-  // PromiseAttributePattern,
-  // FulfilledAttributePattern,
-  // RejectedAttributePattern,
+  PromiseAttributePattern,
+  FulfilledAttributePattern,
+  RejectedAttributePattern,
 } from './resources/template-controllers/promise.js';
 import { AuRender } from './resources/custom-elements/au-render.js';
 import { AuCompose } from './resources/custom-elements/au-compose.js';
@@ -183,9 +183,9 @@ export const PendingTemplateControllerRegistration = PendingTemplateController a
 export const FulfilledTemplateControllerRegistration = FulfilledTemplateController as unknown as IRegistry;
 export const RejectedTemplateControllerRegistration = RejectedTemplateController as unknown as IRegistry;
 // TODO: activate after the attribute parser and/or interpreter such that for `t`, `then` is not picked up.
-// export const PromiseAttributePatternRegistration = PromiseAttributePattern as unknown as IRegistry;
-// export const FulfilledAttributePatternRegistration = FulfilledAttributePattern as unknown as IRegistry;
-// export const RejectedAttributePatternRegistration = RejectedAttributePattern as unknown as IRegistry;
+export const PromiseAttributePatternRegistration = PromiseAttributePattern as unknown as IRegistry;
+export const FulfilledAttributePatternRegistration = FulfilledAttributePattern as unknown as IRegistry;
+export const RejectedAttributePatternRegistration = RejectedAttributePattern as unknown as IRegistry;
 export const AttrBindingBehaviorRegistration = AttrBindingBehavior as unknown as IRegistry;
 export const SelfBindingBehaviorRegistration = SelfBindingBehavior as unknown as IRegistry;
 export const UpdateTriggerBindingBehaviorRegistration = UpdateTriggerBindingBehavior as unknown as IRegistry;
@@ -227,9 +227,9 @@ export const DefaultResources = [
   FulfilledTemplateControllerRegistration,
   RejectedTemplateControllerRegistration,
   // TODO: activate after the attribute parser and/or interpreter such that for `t`, `then` is not picked up.
-  // PromiseAttributePatternRegistration,
-  // FulfilledAttributePatternRegistration,
-  // RejectedAttributePatternRegistration,
+  PromiseAttributePatternRegistration,
+  FulfilledAttributePatternRegistration,
+  RejectedAttributePatternRegistration,
   AttrBindingBehaviorRegistration,
   SelfBindingBehaviorRegistration,
   UpdateTriggerBindingBehaviorRegistration,